Move plugin transport internals into services
Move generic plugin provider transport, declarative provider construction, HTTP subject resolution, and plugin invocation token/server internals from internal/providerhost into service-owned packages.

New interfaces/examples:

- plugins.NewExecutable(ctx, plugins.ExecConfig{Command: command, StaticSpec: plugins.StaticProviderSpec{Name: name}})

- plugins.NewRemote(ctx, client, spec, config, plugins.WithCloser(closer))

- plugins.NewServer(provider)

- plugins.ServeProvider(ctx, provider)

- plugins.NewDeclarativeProvider(manifest, httpClient, plugins.WithDeclarativeConnectionMode(plugins.ConnectionModeUser))

- plugininvoker.NewInvocationTokenManager(secret)

- plugininvoker.NewServer(pluginName, deps, invoker, tokens)

- plugininvoker.InvocationTokenFromContext(ctx)

internal/providerhost now delegates these generic plugin and invocation APIs to services/plugins and services/plugininvoker for compatibility.

Verification:

- go test -p 1 ./services/plugins ./services/plugins/declarative ./services/plugininvoker ./internal/providerhost ./internal/pluginruntime ./internal/bootstrap ./internal/providerdev ./services/providerdev/http ./internal/plugininvocation ./internal/server

- go test ./internal/daemon -run 'TestE2EProviderValidate|TestE2EServePluginOwnedUIWiring|TestE2EHostedHTTPSubjectResolutionUsesAuthorizationAndInheritedInvocation|TestE2EServeStartsWithPluginBoundCacheProvider' -count=1

- golangci-lint run ./...

- git diff --check
