+++ gestaltd/services/plugininvoker/grpc_test.go
@@ -1,4 +1,4 @@
-package providerhost
+package plugininvoker
 
 import (
 	"context"
