+++ gestaltd/services/plugins/http_subject_client.go
@@ -1,4 +1,4 @@
-package providerhost
+package plugins
 
 import (
 	"context"
